Fixed missing _load_settings() for the query keyword.
This has nasty consequences. Due to autovivification, the next code
will create and hashref in the $settings variable:

        if ($settings->{Query}->{log}) {
            $debug = \&_query_debug;
        }

The next check on $settings will return true and the plugin_settings
are not loaded anymore.

The bug pops up when the query keyword is called before other code.
The problem is that the "query" itself doesn't use the settings, and
so the bug doesn't affect "query" but all the other keywords.